requirements.txt
python-dotenv
pandas
numpy
pymongo>=4.6.0
pymongo
certifi
seaborn
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn

